+++ crates/agent/src/agent_configuration/tool_picker.rs
