+++ crates/agent/src/agent_model_selector.rs
